+++ pytorch/src/gen/java/org/bytedeco/pytorch/ConvTranspose1dImpl.java
@@ -53,4 +53,7 @@ public class ConvTranspose1dImpl extends ConvTranspose1dImplBase {
         @Const @ByRef(nullValue = "c10::optional<at::IntArrayRef>(c10::nullopt)") LongArrayRefOptional output_size);
   public native @ByVal Tensor forward(
         @Const @ByRef Tensor input);
+  public native @ByVal Tensor forward(
+        @Const @ByRef Tensor input,
+        @ByRef(nullValue = "c10::optional<at::IntArrayRef>(c10::nullopt)") @Cast({"int64_t*", "c10::ArrayRef<int64_t>", "std::vector<int64_t>&"}) @StdVector long... output_size);
 }